test(llm-api-gateway): move chat allowlist test next to chat route tests
Keep model_uris_test.go for the pure uri matcher unit test and place
TestChatCompletionsModelURIAllowlist with the other chat route tests in
openai_routes_test.go. No test behavior change.
